Factor common methods into LDAUtils
